refactor: renamed /[semster] to [semester] in components directory
