+++ app/(components)/[semester]/LectureListItem.tsx
