+++ app/(components)/[semester]/LectureSearch.tsx
