+++ app/(components)/[semester]/[id]/LectureDescription.tsx
