+++ app/(components)/[semester]/[id]/LectureSchedule.tsx
